:recycle: Move fetch function to FileBrowser Page
